Reevaluation - PaymentRequest Details are stored in the correct places
I realised that although I was being detailed I was not completing the task so I decided to go back on myself and find the simpliest way to get it to store the payment request and ignore the other variables tfor now and then go and improve it once I completed the task.

Currently not done in the best practice however I can improve on the quality once the task has been completeted.